rename footer file to addendum
to avoid confusion with semantic tag footer
